backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
openai==1.51.2
anthropic==0.37.1
cryptography==43.0.1
pydantic==2.9.2
pytest==8.3.3
